+++ client/src/components/Router.tsx
@@ -3,8 +3,8 @@ import { Home } from "../pages/Home";
 // import About from './pages/About';
 // import Login from './pages/Login';
 // import Register from './pages/Register';
-import Register from "./Register/Register";
-import Login from "../pages/login";
+import Register from "../pages/Register/Register";
+import Login from "../pages/Login/Login";
 import { NotFound } from "../errors/NotFound";
 import { Profile } from "../pages/Profile";
 import { Profiles } from "../pages/Profiles";
